ThreadSafety Validations (Rule Creation)
Creating a new Junit4 rule that will run a single test concurrently.  I
think this may have side-effects within the scope of a single Test class,
as how it forces a single single instance of the class to run a specific
test method concurrently. I could not disprove that in initial effort and
wanted to capture the idea.

By default, this rule should have do direct impact on the test
implementations.  Threading is set explicitly to 1 until a System property
is used to override.

EG:
  mvn clean test -DMultithreadRule.concurrency=4

The rule may be added to any Test class which is using the Junit4 test
structures.

** There is always the possibility that